+++ lib/ua_inspector/short_code_map/mobile_browsers.ex
@@ -3,11 +3,9 @@ defmodule UAInspector.ShortCodeMap.MobileBrowsers do
   Mobile Browser Short Code Map.
   """
 
-  @remote_base "https://raw.githubusercontent.com/piwik/device-detector/master"
-
   use UAInspector.ShortCodeMap, [
     file_local:  "short_codes.mobile_browsers.yml",
-    file_remote: "#{ @remote_base }/Parser/Client/Browser.php",
+    file_remote: "Parser/Client/Browser.php",
     var_name:    "mobileOnlyBrowsers",
     var_type:    :list
   ]